requirements.txt
numpy
scipy
pandas
numba
matplotlib
statsmodels
scikit-learn
pyyaml
networkx
pyarrow
tqdm
streamlit
plotly
seaborn
fastapi
uvicorn